# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Shadow Garden API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/guilds:
    get:
      operationId: getGuilds
      tags: [guilds]
      summary: Get all guilds
      parameters:
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: Guilds list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuildsResponse"
  /v1/guilds/{guildId}:
    get:
      operationId: getGuild
      tags: [guilds]
      summary: Get a specific guild by ID
      parameters:
        - in: path
          name: guildId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Guild details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuildDetails"
        "404":
          description: Guild not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
...
    Guild:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        level:
          type: integer
        memberCount:
          type: integer
        ownerName:
          type: string
        createdAt:
          type: integer
      required:
        - id
        - name
        - level
        - memberCount
        - ownerName
    GuildsResponse:
      type: object
      properties:
        guilds:
          type: array
          items:
            $ref: "#/components/schemas/Guild"
      required:
        - guilds
    GuildDetails:
      type: object
      properties:
        guild:
          $ref: "#/components/schemas/Guild"
        members:
          type: array
          items:
            type: object
            properties:
              userId:
                type: string
              name:
                type: string
              level:
                type: integer
              joinedAt:
                type: integer
            required:
              - userId
              - name
              - level
              - joinedAt
      required:
        - guild
        - members
...